add aria-label for accessibility
In order to improve accessibility compliancy, an aria-label need to be added.
"Provide a valid label for form fields" warning from https://www.webaccessibility.com test.
When on-screen labels are present, they must be programmatically associated with form fields. When on-screen labels are not present, form fields must be given an accessible label.
When form fields do not have a programmatic (accessible) label, assistive technologies may incorrectly render the label or provide no label at all to users. When labels are not present or are incorrect, users of assistive technologies may not be able to complete a form.

The HTML5 specification has a new form field attribute called <code>placeholder</code>. This represents a label or hint, such as a word or short phrase, that is assigned to a form field such as an input field. The label or hint appears within the form field and goes away when users start typing. When the <code>placeholder</code> attribute is used, the label or hint may not be detected by assistive technologies. Therefore, developers should provide off-screen text in the <code>label</code> element of the form field or provide a <code>title</code attribute or use ARIA to provide an accessible name for the form field.